+++ packages/bridge-controller/src/utils/metrics/constants.ts
@@ -19,7 +19,8 @@ export enum UnifiedSwapBridgeEventName {
   AllQuotesSorted = `${UNIFIED_SWAP_BRIDGE_EVENT_CATEGORY} All Quotes Sorted`,
   QuoteSelected = `${UNIFIED_SWAP_BRIDGE_EVENT_CATEGORY} Quote Selected`,
   AssetDetailTooltipClicked = `${UNIFIED_SWAP_BRIDGE_EVENT_CATEGORY} Asset Detail Tooltip Clicked`,
-  ResponseValidationFailure = `${UNIFIED_SWAP_BRIDGE_EVENT_CATEGORY} Response Validation Failure`,
+  QuotesValidationFailed = `${UNIFIED_SWAP_BRIDGE_EVENT_CATEGORY} Quotes Failed Validation`,
+  StatusValidationFailed = `${UNIFIED_SWAP_BRIDGE_EVENT_CATEGORY} Status Failed Validation`,
 }
 
 export enum AbortReason {