+++ packages/bridge-status-controller/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Publish `UnifiedSwapBridgeEventName.ResponseValidationFailure` event for invalid getTxStatus responses ([#6362](https://github.com/MetaMask/core/pull/6362))
+- Publish `StatusValidationFailed` event for invalid getTxStatus responses ([#6362](https://github.com/MetaMask/core/pull/6362))
 
 ## [40.1.0]